move type definition into models files
