+++ src/app/backend/query-execution.ts
@@ -1,6 +1,5 @@
 "use server";
-import https from "https";
-import { Bundle, FhirResource, Patient } from "fhir/r4";
+import { FhirResource, Patient } from "fhir/r4";
 
 import { isFhirResource } from "../shared/constants";
 
@@ -12,161 +11,10 @@ import { auditable } from "./auditLogs/decorator";
 import type { QueryDataColumn } from "../(pages)/queryBuilding/utils";
 import type {
   QueryResponse,
-  APIQueryResponse,
   PatientDiscoveryRequest,
   PatientRecordsRequest,
   FullPatientRequest,
 } from "../models/entities/query";
-
-/**
- * Create a FHIR Bundle from the query response.
- * @param queryResponse - The response object to store the results.
- * @returns - The FHIR Bundle of queried data.
- */
-export async function createBundle(
-  queryResponse: QueryResponse,
-): Promise<APIQueryResponse> {
-  const bundle: Bundle = {
-    resourceType: "Bundle",
-    type: "searchset",
-    total: 0,
-    entry: [],
-  };
-
-  Object.entries(queryResponse).forEach(([_, resources]) => {
-    if (Array.isArray(resources)) {
-      resources.forEach((resource) => {
-        bundle.entry?.push({ resource });
-        bundle.total = (bundle.total || 0) + 1;
-      });
-    }
-  });
-
-  return bundle;
-}
-
-/**
- * Tests a connection to a FHIR server from the backend to avoid CORS issues
- * @param url - The URL of the FHIR server to test
- * @param bearerToken - Optional bearer token for authentication
- * @returns Object indicating success/failure and any error messages
- */
-export async function testFhirServerConnection(
-  url: string,
-  bearerToken?: string,
-) {
-  try {
-    const baseUrl = url.replace(/\/$/, "");
-    const searchParams = new URLSearchParams({
-      given: "Hyper",
-      family: "Unlucky",
-      birthdate: "1975-12-06",
-      identifier: "8692756",
-    });
-    const patientSearchUrl = `${baseUrl}/Patient?${searchParams.toString()}`;
-
-    const headers: HeadersInit = {
-      Accept: "application/fhir+json",
-    };
-
-    if (bearerToken) {
-      headers.Authorization = `Bearer ${bearerToken}`;
-    }
-
-    const httpsAgent = new https.Agent({
-      rejectUnauthorized: false,
-    });
-
-    try {
-      const response = await fetch(patientSearchUrl, {
-        method: "GET",
-        headers,
-        // @ts-ignore - Node's fetch types don't include agent, but it works
-        agent: httpsAgent,
-      });
-
-      const responseText = await response.text(); // Get raw response text
-
-      if (response.ok) {
-        try {
-          const data = JSON.parse(responseText);
-
-          if (data.resourceType === "Bundle" && data.type === "searchset") {
-            return { success: true };
-          } else {
-            console.log(
-              "Invalid response structure. Expected Bundle/searchset, got:",
-              {
-                resourceType: data.resourceType,
-                type: data.type,
-              },
-            );
-            return {
-              success: false,
-              error:
-                "Invalid FHIR server response: Server did not return a valid search Bundle",
-            };
-          }
-        } catch (parseError) {
-          console.error("Error parsing JSON response:", parseError);
-          return {
-            success: false,
-            error: "Failed to parse server response as JSON",
-          };
-        }
-      } else {
-        let errorMessage: string;
-        switch (response.status) {
-          case 401:
-            errorMessage =
-              "Connection failed: Authentication required. Please check your credentials.";
-            break;
-          case 403:
-            errorMessage =
-              "Connection failed: Access forbidden. You do not have permission to access this FHIR server.";
-            break;
-          case 404:
-            errorMessage =
-              "Connection failed: The FHIR server endpoint was not found. Please verify the URL.";
-            break;
-          case 408:
-            errorMessage =
-              "Connection failed: The request timed out. The FHIR server took too long to respond.";
-            break;
-          case 500:
-            errorMessage =
-              "Connection failed: Internal server error. The FHIR server encountered an unexpected condition.";
-            break;
-          case 502:
-            errorMessage =
-              "Connection failed: Bad gateway. The FHIR server received an invalid response from upstream.";
-            break;
-          case 503:
-            errorMessage =
-              "Connection failed: The FHIR server is temporarily unavailable or under maintenance.";
-            break;
-          case 504:
-            errorMessage =
-              "Connection failed: Gateway timeout. The upstream server did not respond in time.";
-            break;
-          default:
-            errorMessage = `Connection failed: The FHIR server returned an error. (${response.status} ${response.statusText})`;
-        }
-        return { success: false, error: errorMessage };
-      }
-    } catch (fetchError) {
-      console.error("Fetch error:", fetchError);
-      throw fetchError; // Re-throw to be caught by outer try-catch
-    }
-  } catch (error) {
-    console.error("Overall error in testFhirServerConnection:", error);
-    return {
-      success: false,
-      error:
-        "Connection failed: Unable to reach the FHIR server. Please check if the URL is correct and the server is accessible.",
-    };
-  }
-}
 
 class QueryService {
   /**
